parameters: Allow 'latest' as a valid release value
This is a follow-up of 582278d11d94e1efb72acb0bbec0cd0b0dd3bcee which
introduced support to check for correct release version patterns. This
only allows semvar related values.
However, the `package` command uses the `latest` version by
default. This means that the command `qgis-plugin-ci package` now
fails because `latest` pattern is not allowed.

This issue is fixed by adding `^latest$` to the allowed patterns in
version check.